+++ tests/ChebyshevSharp.Tests/EvaluationValidationTests.cs
@@ -66,6 +66,19 @@ public class EvaluationValidationTests
     }
 
     [Fact]
+    public void Approximation_eval_rejects_out_of_domain_point()
+    {
+        var approx = BuildApproximation();
+
+        Assert.Throws<ArgumentOutOfRangeException>(() =>
+            approx.Eval(new[] { -1.1, 0.0 }, new[] { 0, 0 }));
+        Assert.Throws<ArgumentOutOfRangeException>(() =>
+            approx.VectorizedEval(new[] { 0.0, 1.1 }, new[] { 0, 0 }));
+        Assert.Throws<ArgumentOutOfRangeException>(() =>
+            approx.VectorizedEvalMulti(new[] { 0.0, -1.1 }, new[] { new[] { 0, 0 } }));
+    }
+
+    [Fact]
     public void Approximation_batch_eval_rejects_non_finite_point()
     {
         var approx = BuildApproximation();
@@ -73,6 +86,17 @@ public class EvaluationValidationTests
         Assert.Throws<ArgumentException>(() =>
             approx.VectorizedEvalBatch(
                 new[] { new[] { 0.0, 0.0 }, new[] { double.NaN, 0.0 } },
+                new[] { 0, 0 }));
+    }
+
+    [Fact]
+    public void Approximation_batch_eval_rejects_out_of_domain_point()
+    {
+        var approx = BuildApproximation();
+
+        Assert.Throws<ArgumentOutOfRangeException>(() =>
+            approx.VectorizedEvalBatch(
+                new[] { new[] { 0.0, 0.0 }, new[] { 1.1, 0.0 } },
                 new[] { 0, 0 }));
     }
 
@@ -103,6 +127,17 @@ public class EvaluationValidationTests
     }
 
     [Fact]
+    public void Spline_eval_rejects_out_of_domain_point_before_piece_routing()
+    {
+        var spline = BuildSpline();
+
+        Assert.Throws<ArgumentOutOfRangeException>(() =>
+            spline.Eval(new[] { -1.1, 0.0 }, new[] { 0, 0 }));
+        Assert.Throws<ArgumentOutOfRangeException>(() =>
+            spline.EvalMulti(new[] { 0.0, 1.1 }, new[] { new[] { 0, 0 } }));
+    }
+
+    [Fact]
     public void Spline_batch_eval_rejects_non_finite_point_before_piece_routing()
     {
         var spline = BuildSpline();
@@ -110,6 +145,17 @@ public class EvaluationValidationTests
         Assert.Throws<ArgumentException>(() =>
             spline.EvalBatch(
                 new[] { new[] { 0.0, 0.0 }, new[] { double.NegativeInfinity, 0.0 } },
+                new[] { 0, 0 }));
+    }
+
+    [Fact]
+    public void Spline_batch_eval_rejects_out_of_domain_point_before_piece_routing()
+    {
+        var spline = BuildSpline();
+
+        Assert.Throws<ArgumentOutOfRangeException>(() =>
+            spline.EvalBatch(
+                new[] { new[] { 0.0, 0.0 }, new[] { 0.0, -1.1 } },
                 new[] { 0, 0 }));
     }
 
@@ -125,6 +171,32 @@ public class EvaluationValidationTests
     }
 
     [Fact]
+    public void Slider_eval_rejects_out_of_domain_point()
+    {
+        var slider = BuildSlider();
+
+        Assert.Throws<ArgumentOutOfRangeException>(() =>
+            slider.Eval(new[] { -1.1, 0.0 }, new[] { 0, 0 }));
+        Assert.Throws<ArgumentOutOfRangeException>(() =>
+            slider.EvalMulti(new[] { 0.0, 1.1 }, new[] { new[] { 0, 0 } }));
+    }
+
+    [Fact]
+    public void Slider_eval_multi_rejects_unbuilt_before_domain_validation()
+    {
+        var slider = new ChebyshevSlider(
+            (p, _) => p[0] + p[1],
+            numDimensions: 2,
+            domain: new[] { new[] { -1.0, 1.0 }, new[] { -1.0, 1.0 } },
+            nNodes: new[] { 5, 5 },
+            partition: new[] { new[] { 0 }, new[] { 1 } },
+            pivotPoint: new[] { 0.0, 0.0 });
+
+        Assert.Throws<InvalidOperationException>(() =>
+            slider.EvalMulti(new[] { 0.0, 1.1 }, new[] { new[] { 0, 0 } }));
+    }
+
+    [Fact]
     public void TensorTrain_eval_rejects_non_finite_point()
     {
         var tt = BuildTensorTrain();
@@ -136,12 +208,49 @@ public class EvaluationValidationTests
     }
 
     [Fact]
+    public void TensorTrain_eval_rejects_out_of_domain_point()
+    {
+        var tt = BuildTensorTrain();
+
+        Assert.Throws<ArgumentOutOfRangeException>(() =>
+            tt.Eval(new[] { -1.1, 0.0 }));
+        Assert.Throws<ArgumentOutOfRangeException>(() =>
+            tt.EvalMulti(new[] { 0.0, 1.1 }, new[] { new[] { 0, 0 } }));
+    }
+
+    [Fact]
+    public void TensorTrain_eval_rejects_out_of_domain_user_point_after_reorder()
+    {
+        static double F(double[] p) => p[0] + p[1];
+        var tt = new ChebyshevTT(
+            F,
+            numDimensions: 2,
+            domain: new[] { new[] { 0.0, 1.0 }, new[] { 10.0, 20.0 } },
+            nNodes: new[] { 5, 5 },
+            maxRank: 4);
+        tt.Build(verbose: false, method: "svd");
+        var reord = tt.Reorder(new[] { 1, 0 }, maxRank: 4, tolerance: 1e-12);
+
+        Assert.Throws<ArgumentOutOfRangeException>(() =>
+            reord.Eval(new[] { 15.0, 0.5 }));
+    }
+
+    [Fact]
     public void TensorTrain_batch_eval_rejects_non_finite_point()
     {
         var tt = BuildTensorTrain();
         var points = new double[,] { { 0.0, 0.0 }, { double.NegativeInfinity, 0.0 } };
 
         Assert.Throws<ArgumentException>(() => tt.EvalBatch(points));
+    }
+
+    [Fact]
+    public void TensorTrain_batch_eval_rejects_out_of_domain_point()
+    {
+        var tt = BuildTensorTrain();
+        var points = new double[,] { { 0.0, 0.0 }, { 1.1, 0.0 } };
+
+        Assert.Throws<ArgumentOutOfRangeException>(() => tt.EvalBatch(points));
     }
 
     [Fact]